style: Place Detail 폰트 weight, size 변경 & 파일 위치 이동
